+++ src/test/suite/filterText.test.ts
@@ -1,8 +1,9 @@
 import * as assert from 'assert';
-import { before, after, afterEach, describe } from 'mocha';
-import { closeTextEditor, sleep, getDocumentText, createNewEditor, selectAllText, closeAllEditors } from '../../modules/helpers';
+import { before, after, describe } from 'mocha';
+import { sleep, getDocumentText, createNewEditor, selectAllText, closeAllEditors, findLinesMatchingRegEx } from '../../modules/helpers';
 import { removeDuplicateLines, removeEmptyLines } from '../../modules/filterText';
 import * as os from 'os';
+import { ConfigurationTarget, window, workspace } from 'vscode';
 
 
 suite("filterText", () => {
@@ -21,7 +22,7 @@ suite("filterText", () => {
         const eol = os.EOL;
         const testEditorText = `Fehfomda pemup mihjeb${eol}${eol}uvonono nelvojpo wokragsi geligab${eol}${eol}${eol}pokacan repafme racje ut alhacov${eol}${eol}Hireme gahze${eol}${eol}${eol}${eol}${eol}pi zo iro becago vekabo${eol}luihait abe zukuv gof tususho${eol}${eol}${eol}${eol}`;
         const testEditorTextRedundantExpected = `Fehfomda pemup mihjeb${eol}${eol}uvonono nelvojpo wokragsi geligab${eol}${eol}pokacan repafme racje ut alhacov${eol}${eol}Hireme gahze${eol}${eol}pi zo iro becago vekabo${eol}luihait abe zukuv gof tususho${eol}${eol}`;
-        const testEditorTextAllExpected = `Fehfomda pemup mihjeb${eol}uvonono nelvojpo wokragsi geligab${eol}pokacan repafme racje ut alhacov${eol}Hireme gahze${eol}pi zo iro becago vekabo${eol}luihait abe zukuv gof tususho${eol}${eol}`;
+        const testEditorTextAllExpected = `Fehfomda pemup mihjeb${eol}uvonono nelvojpo wokragsi geligab${eol}pokacan repafme racje ut alhacov${eol}Hireme gahze${eol}pi zo iro becago vekabo${eol}luihait abe zukuv gof tususho${eol}`;
 
         let tests = [
             { docText: testEditorText, redundantOnly: true, expected: testEditorTextRedundantExpected },
@@ -31,7 +32,7 @@ suite("filterText", () => {
         tests.forEach(function (t) {
             let itRedundant;
             t.redundantOnly ? itRedundant = "redundant" : itRedundant = "all";
-            test("Filter " + itRedundant + " empty lines", async () => {
+            test("Remove " + itRedundant + " empty lines", async () => {
                 await createNewEditor(testEditorText);
                 await removeEmptyLines(t.redundantOnly);
                 await sleep(500);
@@ -69,6 +70,49 @@ suite("filterText", () => {
 
                 let text = getDocumentText();
                 assert.deepStrictEqual(text, t.expected);
+            });
+        });
+    });
+
+    describe("Filter lines", () => {
+        const eol = os.EOL;
+        const textToFilter = `pippo${eol}pippo${eol}${eol}paperino${eol}paperino pippo${eol}${eol}${eol}pippo${eol}paperino${eol}pippo paperino${eol}${eol}paperino${eol}paperino${eol}${eol}paperino${eol}`;
+        const regExpExpectedResult = `paperino${eol}paperino pippo${eol}paperino${eol}pippo paperino${eol}paperino${eol}paperino${eol}paperino`;
+
+        let tests = [
+            { textToFilter: textToFilter, regExpString: "/.*paperino.*/gm", expected: regExpExpectedResult },
+            // { textToFilter: textToFilter, regExpString: "paperino", expected: undefined }
+        ];
+
+        tests.forEach(function (t) {
+            test("Filter text with " + t.regExpString, async () => {
+                await createNewEditor(t.textToFilter);
+                let result = findLinesMatchingRegEx(t.regExpString);
+                await createNewEditor(result?.join(eol).toString()); // TODO: convertArrayToText
+                await sleep(500);
+
+                let text = String(getDocumentText());
+                assert.deepStrictEqual(text, t.expected);
+            });
+
+            tests = [
+                // { textToFilter: textToFilter, regExpString: "/.*paperino.*/gm", expected: regExpExpectedResult },
+                { textToFilter: textToFilter, regExpString: "paperino", expected: regExpExpectedResult }
+            ];
+
+            test("Filter text with " + t.regExpString, async () => {
+                let config = workspace.getConfiguration("tt", window.activeTextEditor?.document);
+                await config.update("filtersUseRegularExpressions", false, ConfigurationTarget.Global);
+
+                await createNewEditor(t.textToFilter);
+                let result = findLinesMatchingRegEx(t.regExpString);
+                await createNewEditor(result?.join(eol).toString()); // TODO: convertArrayToText
+                await sleep(500);
+
+                let text = String(getDocumentText());
+                assert.deepStrictEqual(text, t.expected);
+
+                await config.update("filtersUseRegularExpressions", undefined, ConfigurationTarget.Global);
             });
         });
     });